delete unused scheduler
